requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
pyarrow>=14.0.0
google-cloud-storage>=2.14.0
google-cloud-bigquery>=3.25.0
google-cloud-bigquery-storage>=2.25.0
db-dtypes>=1.2.0
pandas-gbq>=0.22.0
graphviz>=0.20.3
# Optional:
# dbt-bigquery>=1.8.0
